edit functions

+++ src/string/s21_memchr.c
@@ -7,11 +7,12 @@
 void *s21_memchr(const void *str, int c, s21_size_t n) {
 
   char *ch = (char *)str;
-  s21_size_t i = (int)n + 1;
+  s21_size_t i = n + 1; // (+ '\0')
   while (i > 0) {
     if (*ch != (char)c) {
       ch = (char *)str++;
-    } else break;
+    } else
+      break;
     i--;
   }
   if (i == 0) {
